Move the dynimage::resize_dimensions function to math::utils
Note that math::utils::resize_dimensions() is crate public.
We don't want to expose it outside of this crate in its current form.

Relates to https://github.com/image-rs/image/issues/1354
