+++ src/math/mod.rs
@@ -4,3 +4,4 @@ pub mod utils;
 
 mod rect;
 pub use self::rect::Rect;
+pub(crate) use self::utils::resize_dimensions;
